+++ src/main/java/lotto/common/LottoConfiguration.java
@@ -2,9 +2,7 @@ package lotto.common;
 
 public class LottoConfiguration {
     static public final Integer MIN_VALUE = 1;
-
     static public final Integer MAX_VALUE = 45;
-
     static public final Integer LOTTO_COUNT = 6;
     private LottoConfiguration() {}
 }
