+++ src/main/java/lotto/view/console/StandardConsole.java
@@ -1,4 +1,6 @@
-package lotto.views;
+package lotto.view.console;
+
+import lotto.view.console.Console;
 
 import java.util.Scanner;
